refactor: use polymorphism for inner nodes
While the tagged pointer was nice, it led to awkward cases where we had
to assert that functions where never called with leaf nodes. There was
no way in the type system to specify just a leaf or just an inner node.

Instead, at the cost of an extra byte per inner node, we can use the
tagged pointer to differentiate between inner and leaf nodes but then
use a tag in the inner node header to differentiate between the
different types of inner nodes.

This allows us to define all the inner node operations on the inner node
itself and not allow these operations to be called on leaves, enforced
at compile time.

A possible alternative would be to support nested tagged pointers, which
seems doable, but I'll save that for another day.